fix(users): update People nav redirects and add usePeopleLists hook (WIP)
- Fixed `People` component so ProfileStats nav redirects to `people/followers` and `people/following`.
- Created `usePeopleLists` hook to handle pure list selection logic and placeholders (no data fetching or search).
