+++ src/hooks/users/index.js
@@ -1,8 +1,9 @@
-export { useUsers } from "./useUsers";
-export { useTopFollowers } from "./useTopFollowers";
 export { useFollowers } from "./useFollowers";
 export { useFollowing } from "./useFollowing";
-export { useUserConnections } from "./useUserConnections";
-export { useProfile } from "./useProfile";
 export { useOnlineStatus } from "./useOnlineStatus";
+export { usePeopleLists } from "./usePeopleLists";
+export { useProfile } from "./useProfile";
 export { useSearch } from "./useSearch";
+export { useTopFollowers } from "./useTopFollowers";
+export { useUserConnections } from "./useUserConnections";
+export { useUsers } from "./useUsers";
